pyproject.toml
[tool.poetry]
name = "ingame"
version = "0.1.0"
description = ""
authors = ["Mikhail Tergoev <57610802+castro-fidel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
PySide6 = "^6.7.0"
PySide6-Essentials = "^6.7.0"
PySide6-Addons = "^6.7.0"
shiboken6 = "^6.7.0"
pyqtgraph = "^0.13.6"
requests = "^2.31.0"
desktop-parser = "^0.1.1"
pygame = "^2.5.2"
python-steam-api = "^2.0"
python-steamgriddb = "^1.0.5"
platformdirs = "^4.2.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
setuptools = "^70.0.0"

[tool.poetry.scripts]
main = 'ingame.main:main'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
